regularization algorithm reorganization
The original regularization algorithm took at top down recursive approach regularizations were identified by in the top nodes and then efficiently
applied down their child nodes

However, this type of regularization fails when there are two atoms only distinguishable by their index.

Example:  Suppose the bonds to these atoms are entirely unspecified and that reactions exist where both of them are single bonds and where one is single and one is double.
During regularization dimension identification the single bond choice becomes a regularization dimension for both because setting either bond to a single bond does not impact which reactions are matched.  The top down regularization algorithm sees these as regularization dimensions and tried to set all of the bond orders to single for both bonds causing it to no longer match the reactions that have the double bond there.

The new algorithm regularizes from the bottom up.  At nodes with no children all graphically indistinguishable nodes are not regularized while (for the case of simpleRegularization) all other nodes are fully regularized.  For nodes with children (in the case of simpleRegularization) full regularization is attempted, but ignored if this causes the template to not match any of its children.